refactor : JPA 리팩토링
