+++ src/main/java/gift/entity/Product.java
@@ -17,7 +17,7 @@ public class Product {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private long id;
+  private Long id;
 
   @NotBlank(message = "이름을 입력해야 합니다.")
   @Size(max = 15, message = "이름은 공백을 포함하여 최대 15자까지 입력할 수 있습니다.")
